// src/games/keybr/src/widget/components/box/Box.types.ts
import type { JSX } from "@solidjs/web";
import type { ValidComponent } from "@solidjs/web";
import { type ClassName } from "../types.ts";
export type BoxDirection = "row" | "column";
export type BoxWrap = "wrap" | "wrap-reverse" | "nowrap";
export type BoxJustifyContent = "start" | "end" | "center" | "space-between" | "space-around";
export type BoxAlignItems = "start" | "end" | "center" | "baseline" | "stretch";
export type BoxAlignContent = "start" | "end" | "center" | "space-between" | "space-around" | "stretch";
export type BoxProps = {
    readonly direction?: BoxDirection;
    readonly wrap?: BoxWrap;
    readonly justifyContent?: BoxJustifyContent;
    readonly alignItems?: BoxAlignItems;
    readonly alignContent?: BoxAlignContent;
    readonly as?: ValidComponent;
    readonly className?: ClassName;
    readonly id?: string;
    readonly title?: string;
    readonly children?: JSX.Element;
};
export type HBoxProps = Omit<BoxProps, "direction">;
export type VBoxProps = Omit<BoxProps, "direction">;